refactor: extract runner types into runner/types.ts
